feat: added Proguard rules for gms Ads
When the app is not using the Play Services Ads lib(as it increases the app size to 1mb) using Proguard removes the gms advertising ID classes and the SDK is not able to fetch the GAID in for any event. Adding the rules in Proguard prevents this from occurring.