Replaced references to MainActivity with Launcher
Added more context around why it's better to use a LauncherActivity instead of the MainActivity.  Since the MainActivity typically can be navigated to during the in-app UX, the LauncherActivity avoids the scenario where the user would be routed again if they navigated back to the home screen.  cc: @aaaronlopez @ttantawy @Sarkar